Sanitized environment variables, in particular LD_LIBRARY_PATH Added extrae-test-dyninst tool to check DynInst functionality

+++ src/launcher/dyninst/cpp_utils.C
@@ -32,7 +32,7 @@
 #include "cpp_utils.h"
 
 void split(vector<string> &tokens, const string &text, char sep) {
-    int start = 0, end = 0; 
+    size_t start = 0, end = 0; 
     while ((end = text.find(sep, start)) != string::npos) {
         tokens.push_back(text.substr(start, end - start));
         start = end + 1;
